SLING-5228 - Remove loginAdministrative() usage from org.apache.sling.installer.provider.jcr
Since the bundle is now using loginService instead of loginAdminstrative
and the commons.testing bundle does not provide support for that I have
moved the tests to use the Sling mocks.

This allows the removal of large chunks of test code. I also took the
opportunity to consolidate more tests under the same superclass where it
made sense.

As a side effect, the code now requires Java 7 but that should be ok.

git-svn-id: https://svn.apache.org/repos/asf/sling/trunk@1771170 13f79535-47bb-0310-9956-ffa450edef68